Turn Matrix into a struct
Zig doesn't have distinct types, which means that 'const A = u32' and
`const B = u32' are one and the same type. This affects type safety,
and also makes it impossible for us to give methods to such a type. In
other words, we can't do what we would do in Go.

The only downside of this change is that the Matrix type is no longer
directly indexable; we have to access the 'data' field first. However,
this change cleans up the API, and gets rid of the matrix.Matrix
stutter.